solver1

+++ src/api/java/cvc5/Solver.java
@@ -821,7 +821,7 @@ public class Solver implements IPointer
    * @param val the value of the constant
    * @return a constant of sort Integer
    */
-  public Term mkReal(int val)
+  public Term mkReal(long val)
   {
     long termPointer = mkRealValue(pointer, val);
     return new Term(this, termPointer);
